db connection related problems are fixed.

+++ Torpedo/src/etri/sdn/controller/module/firewall/OFMFirewall.java
@@ -57,7 +57,12 @@ import etri.sdn.controller.util.Logger;
  */
 public class OFMFirewall extends OFModule implements IFirewallService
 {
-
+	/**
+	 * if this value is false, the firewall module works 
+	 * without persistent database. 
+	 */
+	public boolean dbSupported = true;
+	
 	private FirewallStorage firewallStorage;
 	private OFMStorageManager storageInstance;
 
@@ -512,12 +517,19 @@ public class OFMFirewall extends OFModule implements IFirewallService
 	@Override
 	protected void initialize() {
 
-		TorpedoProperties conf = TorpedoProperties.loadConfiguration();
-
 		this.firewallStorage = new FirewallStorage (this, "FirewallStorage");
-		this.storageInstance = (OFMStorageManager) getModule(IStorageService.class);
-		this.dbName = conf.getString("storage-default-db");
-		this.collectionName = firewallStorage.getName();
+
+		if (dbSupported) {
+			TorpedoProperties conf = TorpedoProperties.loadConfiguration();
+			this.storageInstance = (OFMStorageManager) getModule(IStorageService.class);
+			this.dbName = conf.getString("storage-default-db");
+			this.collectionName = firewallStorage.getName();
+			
+			//check the DB connection, and start without DB if connection failed.
+			if (!storageInstance.isConnected()) {
+				dbSupported = false;
+			}
+		}
 
 		this.protocol = getController().getProtocol();
 
@@ -536,8 +548,10 @@ public class OFMFirewall extends OFModule implements IFirewallService
 				});
 
 		// Read rules
-		synchronized (rules) {
-			this.rules = readRulesFromStorage();
+		if (dbSupported) {
+			synchronized (rules) {
+				this.rules = readRulesFromStorage();
+			}
 		}
 	}
 
@@ -813,7 +827,9 @@ public class OFMFirewall extends OFModule implements IFirewallService
 		entry.put(COLUMN_ACTION, Integer.toString(rule.action.ordinal()));
 
 		firewallStorage.getFirewallEntryTable().insertFirewallEntry(Integer.toString(rule.ruleid), entry);
-		firewallStorage.insertDBEntry(storageInstance, dbName, collectionName, entry);
+		if (dbSupported) {
+			firewallStorage.insertDBEntry(storageInstance, dbName, collectionName, entry);
+		}
 
 		purgeMatchingRulesFromSwitches(rule);
 	}
@@ -821,7 +837,9 @@ public class OFMFirewall extends OFModule implements IFirewallService
 	@Override
 	public void deleteRule(int ruleid) {
 		firewallStorage.getFirewallEntryTable().deleteFirewallEntry(Integer.toString(ruleid));
-		firewallStorage.deleteDBEntry(storageInstance, dbName, collectionName, ruleid);
+		if (dbSupported) {
+			firewallStorage.deleteDBEntry(storageInstance, dbName, collectionName, ruleid);
+		}
 
 		synchronized ( this.rules ) {
 			Iterator<FirewallRule> iter = this.rules.iterator();
@@ -843,7 +861,9 @@ public class OFMFirewall extends OFModule implements IFirewallService
 		synchronized ( this.rules ) {
 			for ( FirewallRule rule : this.rules ) {
 				firewallStorage.getFirewallEntryTable().deleteFirewallEntry(Integer.toString(rule.ruleid));
-				firewallStorage.deleteDBEntry(storageInstance, dbName, collectionName, rule.ruleid);
+				if (dbSupported) {
+					firewallStorage.deleteDBEntry(storageInstance, dbName, collectionName, rule.ruleid);
+				}
 
 				// found the rule, now remove it
 				purgeMatchingRulesFromSwitches(rule);
